# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "School CRM API Documentation",
    "version": "1.0.0",
    "description": "Complete API documentation for School CRM SaaS Platform",
    "contact": {
      "name": "API Support",
      "email": "support@schoolcrm.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://api.schoolcrm.com/api/v1",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/platform/schools": {
      "get": {
        "summary": "Get all schools",
        "tags": [
          "Platform - Schools"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "TRIAL",
                "SUSPENDED",
                "CANCELLED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of schools"
          }
        }
      },
      "post": {
        "summary": "Create new school",
        "tags": [
          "Platform - Schools"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "schoolName",
                  "subdomain",
                  "adminName",
                  "adminEmail",
                  "adminPhone"
                ],
                "properties": {
                  "schoolName": {
                    "type": "string"
                  },
                  "subdomain": {
                    "type": "string"
                  },
                  "adminName": {
                    "type": "string"
                  },
                  "adminEmail": {
                    "type": "string"
                  },
                  "adminPhone": {
                    "type": "string"
                  },
                  "subscriptionPlanId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "School created"
          }
        }
      }
    },
    "/platform/schools/{id}": {
      "get": {
        "summary": "Get school by ID",
        "tags": [
          "Platform - Schools"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "School details"
          },
          "404": {
            "description": "School not found"
          }
        }
      },
      "put": {
        "summary": "Update school",
        "tags": [
          "Platform - Schools"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "School updated"
          }
        }
      },
      "delete": {
        "summary": "Delete school",
        "tags": [
          "Platform - Schools"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "School deleted"
          }
        }
      }
    },
...
    "/school/stats": {
      "get": {
        "summary": "Get all stats",
        "tags": [
          "School - Transport"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/school/": {
      "post": {
        "summary": "Create resource",
        "tags": [
          "School - Timetable"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/school/:id": {
      "put": {
        "summary": "Update :id",
        "tags": [
          "School - Timetable"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "summary": "Delete :id",
        "tags": [
          "School - Timetable"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/school/routes": {
      "get": {
        "summary": "Get all routes",
        "tags": [
          "School - Transport"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "summary": "Create routes",
        "tags": [
          "School - Transport"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/school/routes/:id": {
      "put": {
        "summary": "Update routes",
        "tags": [
          "School - Transport"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "summary": "Delete routes",
        "tags": [
          "School - Transport"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/school/users": {
      "get": {
        "summary": "Get all users in school",
        "tags": [
          "School - User Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of users"
          }
        }
      },
      "post": {
        "summary": "Create new user",
        "tags": [
          "School - User Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "fullName",
                  "email",
                  "role",
                  "password"
                ],
                "properties": {
                  "fullName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "SCHOOL_ADMIN",
                      "PRINCIPAL",
                      "TEACHER",
                      "STUDENT",
                      "PARENT",
                      "ACCOUNTANT",
                      "LIBRARIAN",
                      "RECEPTIONIST"
                    ]
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          }
        }
      }
    },
    "/school/users/{id}": {
      "get": {
        "summary": "Get user by ID",
        "tags": [
          "School - User Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details"
          }
        }
      }
    },
...
  },
...
}